Move redux components within directory structure
