+++ src/scripts/Main/redux/epics/authenticateEpics.ts
@@ -1,9 +1,14 @@
+/**
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license http://www.opensource.org/licenses/gpl-2.0.php GNU GPL v2
+ */
+
 import { combineEpics } from "redux-observable";
 import apiv2 from "@core/apiv2";
 import { log, logError } from "@core/utility";
 import "rxjs/add/operator/mergeMap";
 import "rxjs/add/operator/throttleTime";
-import { AUTHENTICATE_AUTHENTICATORS_GET, authenticatorsSet } from "./actions";
+import { AUTHENTICATE_AUTHENTICATORS_GET, authenticatorsSet } from "../actions/authenticateActions";
 
 // Logging helper HOC to make code a little more readable
 const logResponse = prefix => {
