+++ src/scripts/Main/redux/reducers/authenticateReducer.ts
@@ -1,4 +1,9 @@
-import { AUTHENTICATE_AUTHENTICATORS_SET } from "./actions";
+/**
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license http://www.opensource.org/licenses/gpl-2.0.php GNU GPL v2
+ */
+
+import { AUTHENTICATE_AUTHENTICATORS_SET } from "../actions/authenticateActions";
 
 const initialState = {
     authenticators: [],
